Update accessibility specs
Introducing the cookie banner appears to have broken our accessibility
specs, we can only assume that intoducing content above the skip link
makes Axe raise an error.

This is clearly noted in the documetation under
'When to use this component' here:

https://design-system.service.gov.uk/components/skip-link/

To work around this we decided to exclude the skip link in the test, but
could not find a way to configure this for all specs. We've ended up
creating a helper method that encapsulates the exclusion of the skip
link and whilst there, ensures we are testing against WCAG2.2AA.