Change how the conversion date is shown
When a project is created the conversion date may only be provisional,
i.e. not yet agreed.

Previously we made this clear to users by labelling the field
'Provisional conversion date', but we are about to add the ability to
store the 'agreed converision date'.

Ahead of that we change how this field is displayed to include a
provisional tag and 'converting on' label, this will let the field
display either value.

Right now all converting on dates are provisional, collecting the agreed
date will come later.