Add basic collection of confirmed conversion date
This is the first time we are collecting data as part of a task. As this
is a date it is possibly the most complex data type to collect as the
GOVUK pattern splits the day/month/year values into seperate fields.

The basic principle here is that the Task object gets four new
attributes so that the form can be submitted, our Task objects loosly
follow the 'form object' pattern already, passing the submitted values
onto the TaskList for saving to the datebase.

As the task uses `assign_attributes` the three seperate fields are cast
back into a date object and stored.

Because we have to declare the three fields on the Task as attributes,
the task is is never consided complete unless all the fields have values
- which is rarely the case with the extra fields - this means we have to
overide the default `completed?` method to exclude the extra fields.

For the form itself, we have used the GOVUK form builder helpers to
produce a date field widget, to insert our guidance we had to manually
render the hint and guidance to get the desired layout - future work
could subclass `govuk_date_field` for our purposes, but is possibly
only valuable once we have more dates to collects - for now, this is
right enough.