Project supplies appropriate conversion date
Voluntary conversion Projects now have multiple conversion dates, we
first get a provisional date when the project is created and later
confirm the date with all relevant parties.

It makes sense to let the conversion project decide which date is
appropriate so we don't have to.

Here we create a simple ConversionDate class that is returned from
calling `conversion_date` on a conversion project.

The object allows us to indicate if the date is provisional or not and
store the date itself - all this simplifies the interface as I hope you
can see in the project helper, where we act on which date by showing the
users a provisional tag when the date is, well, provisional!

When loaded generic projects we lazy load the task_lists so that the
conversion_date can be calculated.