Remove ProjectPresenter from codebase
We have decided to remove presenters from the project, as we have been
using helpers more organically. Therefore we want to make sure we are
consistent throughout the project and remove presenters where they exist.

I have added the formatting logic into the view, as this felt like the
simplest step forward.