Add basic confirm conversion date validation
As a date that is always the first of the month, we want to only allow
vaules within a range. This is made simpler by our Task models following
a loose 'form object' pattern, where we can validate the month and year.

We have to use a custom validator as we want all errors to be returned
to the 'containing' field, in this case `confirmed_conversion_date` -
this allows the errors to show up correctly on the page.