Alias the `contacts` route to `external_contacts`
The "Contacts" are now "External contacts". We want these contacts to be
available on a route `external_contacts` to differenciate them from the
"internal" contacts we will be adding.

Aloiasing the route in this way is easier than renaming the Contacts model to
ExternalContact. We will not be adding the new Internal contacts as contact
models, instead they are kept as properties of the project. So it makes sense
(for now) to keep the External contacts as `Contact` models, but change the
route with an alias.