Fix: Enable Google analytics helper
Previously the last check performed by the helper was to check the value
of the cookie, the check was for the true object, which a cookie will
not return as cookies store JSON and so must be parsed first.

As with the check for the SENTRY_ENV, we check for a string which fixes
the behavioiur.