(Refactor) Refactor this spec to better reflect roles
This spec is a little hard to read as it had two "users" with similar names.
One appears to be acting as a caseworker and one as an unassigned user.

Refactor this spec slightly to change the user names, `user_1` becomes `user`
and `user_2` can be renamed to `caseworker` without altering any of the specs.

The wording of some of the specs has also been changed to make it clearer
what they are testing.

Additionally, we have added a caseworker context to show that the caseworker
can see projects they are the caseworker on.