Show `assigned_to_regional_caseworker_team` radio buttons on Project form
This is for Voluntary projects only.

Add a pair of radio buttons to the Conversion Project create form, which allow
the user to select Yes or No for `assigned_to_regional_caseworker_team`.

The value is passed through the form to the project and stored in the
attribute `assigned_to_regional_caseworker_team` on the project.

For involuntary projects, the radio buttons are not shown and the attribute
is always stored as `false`.