Set `assigned_to` on a project it is not being assigned to caseworker team
If a voluntary project is being assigned to Regional Caseworker Team, do
not set `assigned_to` on the project, and send the team leader notifications
as usual.

If a voluntary project is not being assigned to Regional Caseworker Team,
set `assigned_to` on the project to the current user.

This does not affect involuntary projects.