Add `check_provisional_conversion_date` to stakeholder kick-off task
The title of the `check_provisional_conversion_date` field on the stakeholder
kick-off task requires a variable.

Therefore (for now) we have added the checkbox partial code "longhand" to the
template, in order to pass in the required variable to the I18n string.

This can no doubt be refactored later into something better?