Refactor `Project.unassigned`
We introduced a `Project.unassigned` scope to surface projects which are
unassigned to a user but assigned to the regional caseworker team. After some
thought we have decided to split this into two scopes, which we can chain
together for the same effect. This then means we can use the two scopes
`Project.unassigned_to_user` and `Project.assigned_to_regional_caseworker_team`
separately.