Allow Team leaders to assign projects from the project list page
For team leaders only, add an "Assign" button to projects on the Unassigned
projects tab. This button allows the Team leader to assign another user to the
project. Once a user is assigned, the Team leader is redirected to the Unassigned
projects list. The project will have vanished from this list as it is now assigned.