Add new `CONTAINER_VNET` value to `config.hosts`
Re: https://ukgovernmentdfe.slack.com/archives/C03AANC6UJF/p1677253180105049

We added a new `/healthcheck` endpoint for monitoring.

Unfortunately when implementing the monitoring, techops encountered an error
because the health probes' IP address is not in the app's allowed hosts:

```
ERROR -- : [ActionDispatch::HostAuthorization::DefaultResponseApp] Blocked host: 172.16.0.126
```

As a solution, techops are going to add a new environment variable to terraform
called `CONTAINER_VNET`. This will be an IP range.

https://guides.rubyonrails.org/configuring.html#actiondispatch-hostauthorization

We can add this to the allowed hosts in the application via
`IPAddr.new(ENV["CONTAINER_VNET"])`