Add Project Openers table
Add a route to a new view at `projects/openers/:month/:year`

This page shows a tabular view of all projects. Eventually the list of projects
will be filtered to projects which have a confirmed opening date of the
supplied year & month, but this has not been implemented yet. So for now this
shows all projects.

If a year or month are not supplied in the URL, the user is routed to a 404
page.

If the year is not in the scope of 2000-2499, or the month is not in the scope
of 1-12, the user is routed to a 404 page.

If no projects fall within the scope of a correctly supplied month and year, an
appropriate informational message will be shown (this is not yet implemented as
we don't currently scope the returned projects).