Add conversion_date to project
The conversion date is a vital attribute to our users and is used to
sort and expose projects in various views.

The is some complexity around how we capture and store this attribute as
we want to expose it at various points:

- during project creation when it is considered 'provisional'
- in an early task when it is 'agreed'
- at any point later when a project 'slips'

We have looked at all the options to model this in the data and keep
things simple and not found anything that does everything.

This possibly speaks to a wider model issue we will have to address,
but right now we need a good enough way to handle this idea of
conversion date.

Having the conversion date makes the attribute readily available in
scopes to sort projects. But makes it difficult to set from a task.

Having the seperate `provisional_conversion_date` and `conversion_date`
lets us sort project by this concept, we know this to be useful, on the
openers list for example.

Later work will cover when a project 'slips' as this has it's own UI and
user journey and can almost be considered distinct