Update how project conversion date is shown
We have the new `conversion_date` attribute on Project that will become
the source of all conversion dates once the date is no longer considered
provisional.

Here we simplify the way the conversion date is shown.

Right now, there is no way for users to set this value, it will only be
set if the conversion date is already agreed for a project, but we will
add that behaviour next.