Voluntary task list syncs conversion date
When a users completes the 'External stakeholder kick off' task, they
tell us what the agreed conversion date is.

Here we sync that value to the `Project.conversion_date`.

Once the `Project.converison_date` is set, we do not want users to be
able to change the date again, we will introduce the behaviour where a
date slips, but that is later work.

We only allow the date to be syncronised once, when the
`Project.conversion_date` is set, it can no longer be modified in this
way.

We have hidden the entire field on the task for now, we may come back
and show it in a disabled state, but that can come later.