Add model for a conversion date history
We will store the previous state of the `Project.conversion_date` each
time it is updated. This change of conversion date is thought of as a
'slip', although I have gone for a more generic `revised_date` for the
column name.

It is important to know why a project slipped, so we link the date
history to a note so users can record why the project slipped as well as
the previous date and the new slipped date.