New converison date history form
This form object sits between the view and model layer so that we can
create all the database objects when adding a new converison date
history.

The process is:

- create the note to capture why the date is changing
- create the history item with the previous and revised date
- update the converison_date attribute of the project

If any one of these operations fails, we do not want orphanned objects,
so the process is wrapped in a transaction that rolls back the objects
unless all are saved successfully.

The form object also allows us to validate we have all the attributes we
need including dependents such as a user id and project id.