Constrain the Project#show routes by UUID
We want to make sure that when a user views a project, they are only passing
a valid UUID. This is to ensure that routes like `/projects/openers` or
`projects/some-accidental-string` are routed to a 404 and not a 500.

Previously, these incorrect routes were routing to the Project#show action
and the application was trying to find a project with that "ID" (e.g an ID of
"openers", not finding the ID and returning a 500 error. We want the application
to return a 404 instead, which makes for a better user experience.