Add a policy around changing the conversion date
The project conversion date goes through three main states:

- provisional
- confirmed
- slipped or delayed

Previously, users could change the conversion date at any time.

It only makes sense to change the conversion date once it is confirmed
as that is the point at which it becomes an unexpected change.

Here we add a simple policy on project to control when a project can
have the conversion date changed outside of it being confirmed.

The policy is used to prevent the acitons and to hide the button that
allows users to change the conversion date.