Use Project.opening_by_month_year scope in project openers table
Now that we have a scope to only view projects opening in a certain month & year
(`Project.opening_by_month_year(month, year)`) we can use it to scope the
tabular view of projects. This view only shows projects with a confirmed
conversion date of the supplied month & year, and a helpful error messages if
there are no projects in that scope.