Sort the projects in the openers table via a service object
The projects in the openers table should be sorted by conditions_met? (true
first) and then establishment name.

Create a service object ProjectsFetcher with a method `sorted_openers` which
fetches the desired projects by month & year, and sorts them first by
all_conditions_met? = true and then by school name.

In order to make this easier to test, we have moved the `includes(:task_list)`
into the `Project.opening_by_month_year` scope - this does not affect
functionality but means we don't have to mock a response to `includes()` in the
mocked Project doubles in the ProjectFetcher spec.