Fetch and show the school address
We know that Academy operational pracitce unit users rely on the address
of the school as part of sending the fundaing agreement letters.

As we fetch most school information from Get information about schools
via the academies API, we want to validate we can fetch the address.

Once we have the address and show it, users can gain confidence that it
is the 'correct' one.

Later we can then export the address in ways that are useful to users,
but that comes later.

Some decisions here:

- went for the address over multiple lines [1]
- wrapped the address in the <address> tag, will worry about this not
  working in older browsers if it is an issue
- added a basic style as the address is render in italics otherwise;
  might be a browser default?
- refactored the project information feature specs into their own
  directory and split them up to aid readability

[1] https://www.gov.uk/guidance/style-guide/a-to-z-of-gov-uk-style#addresses-in-the-uk