Refactor provisional conversion date
This work is the start of a refactoring of how we model the conversion
date.

The conversion date moves through three main states:

provisional > confirmed > changed

It can change multiple times, but it is only ever provisional once. A
chagned date is still considered confirmed as it would have to be
agreed by all parties.

This work is prompted by the difficulty we face building views that rely
on the converison date being one column in the database when it may be
represented by two, `provisional_converison_date` or `conversion_date`.

Here we take a first step in the refactor, apologies to anyone reviewing
this commit as it touches so much - which I think speaks to why this
needs refactoring!

We add a new `conversion_date_provisional` column so we can know from
the database when a date is in that first state.

Where we set `provisional_conversion_date` we also set
`converison_date`, along with `conversion_date_provisional` defaulting
to true, this means the inital project always has a conversion date that
is considered provisional.

For now, we keep `provisional_converison_date` as I do not want to make
this commit any bigger and we need to migrate the values - which will be
tricky.

With this refactor we get two big benefits:

- the conversion date is now one column
- the state of the conversion date (provisional or not) is also a
  database column

Which will make managing views of project far simpler.

The only complexity left is syncing the first confirmed converison
date from the task list to the project conversion date, we haven't
changed this so not gotten any worse. We will have to create a date
history for that change, but we have the callback to do so, this will
happen in a later commit.