Converison date updater service object
I am not sure this is a 'service' object but, it is close enough to live
in the directory!

We are going to have to create a conversion date history when the user
confirms the conversion date for the first time as part of the external
stakeholder kick off task.

The current immplementation is inside the new_date_history_form, here we
extract it to a object for use in other locations.