Remove Caseworker and Team Leader assignment from internal contacts
Due to confusion from users about what 1) a "team leader" actually is, and
2) the difference between a caseworker and the "assigned to" person on a project
we have decided to hide these two fields from users.

The team leader will always be one of two people, and the "assigned to" person
is the field we want users to populate & pay attention to.

Along with removing the two fields from the view, we have repurposed some of
the tests to test the assignment of "assigned_to" still works correctly.