Confirm conversion date action creates history
Previously when the user first confirms the conversion date, we would
set the value of `conversion_date` - now we want to record a converison
date history.

Here we use the new service to do just that. As any conversion date
history must include a User, we make that dependency available by
creating a `user` attribute on `TaskList` sub-classes. You can think of
this as:

> the user who is updateding the task list

We raise an error if a developer uses this feature without setting the
user first.