Repurpose the CaseworkerMailer for project assignees
Now that we no longer assign a Caseworker to a project, we can repurpose
the CaseworkerMailer to send messages to the "assigned_to" person on a project
instead. Rename the mailer & its test, and send a mail when a person is added
as the "assigned_to" person on a project.