Add service and migrate conversions dates
We are working towards a world where there is no
`provisional_converison_date` and need to migrate existing projects to
the new model.

In production right now all Projects have a
`provisional_conversion_date`
and `conversion_date` does not exist so we can run this migration which:

For a project where:

- conversion_date is nil
- conversion_date_provisional is true (the default)

We consider the project provisional and simply copy the date from
provisional_conversion_date to conversion_date

For a all other projects

- conversion_date is set so the conversion date has been agreed

We create a new Conversion::Date history using the
provisional_conversion_date as the previous value and the
conversion_date as the revised one - this would be the same process that
would have happend if the conversion date was agreed with the new model
in place.

With the migration completed successfully, the service can be removed
from the code base.