Rename `caseworker_assigned_at` to `assigned_at`
The `caseworker_assigned_at` attribute on Projects is now redundant, as we are
assigning people to the projects in the `assigned_to` field - and they may not
be caseworkers.

Rename the column `caseworker_assigned_at` to just `assigned_at`
Fill this column whenever anyone is added to the `assigned_to` field on a
project

If any projects are assigned but do not have an `assigned_at` value (e.g. if
they are assigned to an RDO, not a caseworker), backfill the column with the
project's `created_at` date (it's close enough!)