Openers list uses new provisional conversion date
The openers list is very much focused on the idea that it only list
projects that are 'likely' to open.

The two states for this are:

- when the conversion date is agreed
- when the all conditions met task is complete

As we have refactored the converison date, we need to reflect that
change by only including project where the conversion date is not
provisional.