Fix the conversion date migration
When we attempted to run the conversion date migration in development it
failed. This was because any existing project did not get the default
value of true for the new `conversion_date_provisional` meaning the
first if case did not pass and the project fell into the else.

There the code attepmts to create a history but cannot as there is no
value to get a revised date (because it is acually provisional!)

This fixed version:

- changes the first if case to check for a nil converison date, a
  present provisional conversion date and a stake holder kick off task
  that has no value, in production this should mean a project that is
  still provisional
- we use an elsif to confirm the other case has a value for the
  stakeholder kick off task we can use before attempting to do so.
- if neither of these cases is true we do nothing, although that would
  mean a project in an unexpected state.