Add all in progress view
We want to re-renforce the idea that some views are of 'the user's own'
projects and some are 'all projects' - this is ahead of refinements to
our permissions models.

Here we support this concept by adding a view of all projects at
`/projects/all/in-progress`.

Later work will prevent the user who is not 'assigned to' a project from
modifying it.