Prefetch establishment and trust details
On prject index views, views showing multiple projects; It is more
perfromant to fetch the establishment and trust details in bulk.

Worht noting that the pager returns 10 records by default, so anything
that is paged has a 10 record limit.

Here we split the prefetching and use it on all the index views for
projects.

We introduce a new mocker helper that simply allows prefetching of any
establishments and trust - we can do this because we always create the
projects under test, which already have and establishment and trust from
the factory.

We've tried to call this new method something that speaks to its
behaviour:

`mock_pre_fetched_api_responses_for_any_establishment_and_trust`

We also created a sort of alias method, again with a hopefully clearer
name for use when you just want to create new project factories, which
require the mocked API calls but don't need to actually do anything.