Add a regional casework sevices in progress view
As we work towards a logical (and maybe restul-ish?) world it make sense
to build out index views matching specific critera.

Here we want to show project that are handed over to the regional
casework team. This will help the team leads to view projects as they
will not be assignd_to, projects themselves.