Add Regional casework completed projects
This rounds out the collection of four logical views, this one shows all
completed project that have been assigend to the regional casework
services team.

There is a refactor of some locales in here where associated the locale
to for the empty state of a table reflects it's use in the table
partial.