Add a user projects view
This view joins the others to represent the three user facing scopes we
have now:

- all projects
- regional casework services projects
- user projects

This user one will replace the current 'In progress' project view and is
more explicitally the 'users assigned projects'.

Because this is the main view for most users, we don't use a table to
list projects, instead it uses the existing projects list.