Refactor all projects views
We recently introduced the concept of all project views, showing all
projects in the applicaiton that are either:

- in progress
- completed

Here we refactor the controllers into a namespace to thin out the
projects controller.

We strip out the individual policy methods in favour of simply using the
`index?` policy method.

This does lead to some duplication of code with the pre fetching, but we
can probably resolve that later.