Add unassigned projects view to casework serives
This view already exists, but we are working towards a more logical
grouping of views that show multiple projects.

This view is of projects assigned to regional casework services that
have not yet been assigend to a user.

Here we are moving the view into the `regional-casework-services`
namespace and switching to a table view.