Add task list permission
The primary action our users perform is checking off action on a
projects task list.

We want to prevent users who are not assigned to a project from checking
off tasks.

Here we introduce a TaskListPolicy to do just this.

Viewing a task (the form) is allowed for any user, but submission of
the form to change the values is prevented.

Later work will improve how we handle an unathorised user.

We have used the Pudit RSpec DSL to test this:

https://github.com/varvet/pundit#policy-specs