Introduce project update policy
The only time a user who is not a regional delivery officer changes the
Project model is when they 'compelte' a project at the end of the
process.

We only want the user who is `assigned_to` the project to be allowed to
do this.

We also only want incomplete projects to be able to be completed.

Here we introduce a generic `update?` policy to allow this.