Add new policies for new notes and contact
Users should not be adding new contacts and notes to completed projects.

To prevent this we add new policies to the projec policy - this is
becasue at the point we create a new note or contact the only object
that exists it the project to which the note or contact will be
associated.

We can use the policy in the controllers by calling the project policy
specifically.