Introduce a contacts policy
Previously, any users could create a new cotnact at anytime.

We want to prevent users from creating new contacts on completed
projects.

Here we introduce a policy that does just that. It relies on the Project
as creating new instances of a contact requires a parent Project. We use
the new `new_contact?` policy to achieve this when creating new cotnacts
and the contact policy to edit and destroy exsiting contacts.