Updte the note policy
Like contacts, users should not create new notes on completed projects,
we handle this the same way as contacts, by delegating the policy to a
method on the project policy, if the project is completed we do not
allow the note to be created.

Editing and deleting is the same, only as the note exisits, we do not
have to delegate up to the parent project policy.

Becuase the note policy is now dependent on the project being loaded, we
eager load projects when fetching notes.