Bug fix: external stakeholder kick off task
When the projects conversion date is confrimed we do not show the
conversion date fields on the external stakeholder kick off task.

When submitted the task list attempts to save the task expecting all of
the attributes to be present for mass assignment, but because the form
does not show the date fields this fails.

We saw this in production, here is a link to the exception in Sentry:

https://sdd-n7.sentry.io/issues/4001605117/?project=6684508

The solution is tricky, mostly because of our reliance on mass
assignment on the task list.

However as we have this issus in production, this is a proposed qucik
fix.

Once the conversion date is confirmed we add hidden fields to the form
to ensure all attributes are passed along correctly.

At this point, we are not sure what the long term fix will be for
this issue, but we do need this quick fix.