Update notification banner view component
We use this component to render Rails flash messages, we want to also be
able to render the same component with just a messsage, as a
notificaiton informing users when a project is completed and should not
be updated, for example.

Here we extend the component to take a `flashes` object or a simple
mesasage.