Add Process the Sponsored support grant task to Voluntary task list
Add a new task "Process the Sponsored support grant" to the task list for
Voluntary conversions. This task is identical to the "Process conversion grant"
task for involuntary conversions (with the exception that it is optional).

Eventually the involuntary task list will be removed, and this addtition to the
Voluntary task list means we should be able to support all conversion types.