Support different root views
When users sign in they are redirected to the root path, this has
always be the 'in progress' projects list, but that is not the most
useful view for all users.

Here we switch from a hardcoded controller action to a rediect that we
customise based on the current user.

Right now we send team leads to the unassigned project list and everyone
else to the existing 'my in progress projects' view - later we can do
more.