Refactor 'home' page tabs
Now we can customise the page to a user we can show the team leads a
view of all their teams projects, both in progress and compelted along
with the unassigned view.

These now use the new versions of these views.

This also makes the `unassigend?` project policy redundant as we don't
need to use policies here.