Fix: link back to projects
Product review for release 17 found this bug where a regional delivery
officer creates a new projec to be handed over to regional casework
services, upon creating the project the link back to their own projects
view was broken and casued an error.

We fix the link to the users in progress projects.

We also remove the route for `/projects` which has no controller action
and update any places we rely on that route.