Validate new projects URN
The same school should never been converting more that once, we've seen
some issues in production where different users add the same project
resulting in duplicates.

This works prevents the issue by checking that a project that is in
progress does not already exist and if it does, showing a simple error
message.