Introduce the sponsored route
The Sponsored route replaces the old involuntary route, we have taken
steps to stop users adding any new involuntary converions and work will
start soon on stripping out the code.

In the meantime we want to show users the 'route' a project takes which
is now represented by the `sponsor_trust_required` attribute.

If a sponsor trust is required the project is 'Sponsored' if not, it is
'Voluntary'.