Add all in progress projects views
We have the 'all' projects scope and want to divide that down into all
sponsored and all voluntary projects as well.

Here we create a new all in progress controller, views and test to cover
this.

We will follow up with a matching set for completed projects.

We use the all scope because, right now regional casework services do
not handle sponsored conversions so it only makes sense in the all
scope.