In-progress projects does not include unassigned
Our current model of 'in-progress' is that the completed at time is not
set, it feels right to add to this projects that are assigned to a user,
if the project was not assigned who would do the work to make it
in-progress?

Here we address this in the in-progress scope.

We may want to add a view that exposed projects that are neither
in-progress or completed, for now we don't need to worry because:

- for regional caswork services they can already do this with the
  unassigned view
- for regional users, they are auto assigned to the project at creation
  so the risk is minimised

To keep the test suite green, we update the project factories to include
an assigned user - this feels right as the base factory should
represent the most common state. As the project factory is fired a lot,
we set a random email address to pass the unique constraint on the email
column.

To keep the scopes clean we introduce an 'assigned' scope the the
in-progress scope can utilise.