Add basic Application Insights support
We have been asked to follow all the .NET products in the program and
use Microsoft Applcation Insights. Whilst not something the team is
familiar with, we want to try and accomadate where it does not effect
our ability to deliver value or support the live service.

The gem used here is set to archived on Github [1], but may be stable
enough for our needs, we have to test it to find out.

After a small spike, here we have the absolute minimum working version
of this, with basic stats going to Application Insights.

We have set the following options as the defaults:

- buffer size: 500
- send interval: 60

See: https://github.com/microsoft/ApplicationInsights-Ruby/blob/5db6b4ad65262d23f26b678143a4a1fd7939e5c2/lib/application_insights/rack/track_request.rb#L18

We will ship this version and let it run for a period of time to try and
better understand:

- what we can do with Application Insights
- how it aligns with what we are being asked to do AKA compared to other
  .NET products

We only load Application Insights if the key is present and in the
three 'production' environments:

- development
- test
- production

[1] https://github.com/Microsoft/ApplicationInsights-Ruby