Add a conversion date revised from scope
A project conversion date may change over time, going from:

provisional > confirmed > changed

It may be changed any number of times.

We store the conversion date in a column and the history of changes in a
table.

Some of our users want to see which projects were due to convert in a
given month, but have since 'slipped' and will no longer. We understand
this is because they may have prepared work based on the previous
conversion date and need to know not to progress that work now the
conversion date has slipped.

This scope enables us to retrieve projects that match this criteria,
it's a bit complicated so we'll try and go over it here.

The first step is to get all the lastest DateHistory for each project,
we do this with `group` on project_id and `maximum` on the created_at date.
This gets us the latest DateHistory per project, because you can only
group and fetch at the same time, this gives us the set of project_id
and the created_at date, but nothing more.

So we take those two values and create an sql query that fetches all the
DateHistories with all their attributes by matching the project ID and the
created at date, this works because projects can only have one
DateHistory as the latest in it's series by created_at (time is funny
that way!).

With this query built, we pass it into the last query to fetch the
projects that match, again via the project_id on the DateHistory. This
query adds:

- excluding any results where the previous and revised date are the
  same, in this case there has been no change to the date and is likely
  the 'confirmed' date - we expect this.
- where the previous date is in the same month as the one supplied to
  the scope
- where the previous date is the same year as the one supplied to the
  scope.

We build this query as sql so that it can be executed along with the
final join - this increases the performance as we only have to go to the
database twice for the whole operation.

Finally, we sort these by conversion date.

We make heavy use of existing scopes to ensure the set of projects we
are working with is a small as possible, for this reason the resulting
test is actually quite small, as we only have to test the new
dimensions.

Big thank you to @thill-dxw, who helped flesh this out in sql before
taking it to Rails!