View of projects that were going to convert
Conversion project have the concpet of the converison date, the day on
which the school closes and the new academy opens. This is a hugely
important date.

When the conversion date of a project changes various users are
interested in knowing this information. Right now we think this is due
to the current work practices rather than any real need to know this.
But as we need to support a balance between what is happening now and
what will happen once the application is the sole place to come for
information, we want to prove we can handle this use case.

Here we add a view, the 'slipped projects' that given a month and year,
shows the projects that were due to convert in that time but have since
changed.