Add the academy details to project information
Whilst we do not actually fetch the academy details yet, this lays the
ground work for that to have a place to be displayed.

Projects will exist without a academy URN for some time before being
provided with one, so this work handles that case by showing a message
in the appropriate section.