Fix route
As soon as we deployed the new voluntary/sponsored route work we got a
new project added in a state we thought would not exists.

We think this was caused by the language used to phrase the quesiton on
the new project form:

> Is the school joining a Sponsor trust?

rather than:

> Does this project require a sponsor?

Schools can voluntarily joing a sponsor capable trust, but the
converison is not sponsored, i.e. will not recieve additional funding -
in these circumstances the response to the question is: yes - which
results in a sponsored route - but is infact a voluntary one!

This change stops this being the case by relying on the
`directive_academy_order` attribute which, if true, will always mean a
sponsored conversion.

Whether we should retain the sponsor question at all remains to be seen.