Do not collect sponsor trust required
We have decided the link between this attribute and
directive_academy_order is 100% so we do not need to collect this.

There is another question that is subltly different that is:

> is the school converting voluntarily and joining a trust that is
capable of sponsoring, but this is of no value

This work stops asking for the value and instead sets it based on the
value given for `directive_academy_order`.