Remove sponsor_trust_required
As mentioned in 6456b37e0dae7c4a1fb372b88416caf0892d4d86 the concept of
'is a school required to join a sponsor trust and so be a sponsored
conversion' is solely a factor of whether a directive academy order has
been issued or not.

This work removes the sponsored_trust_required from the code.

There is still a subtly different quesiton that we came accross in
production that is:

> is the school voluntarily converting into an academy by joining a
sponsor capable trust, but is not being sponsored

But we are confident this is not something valuable to the service.