Add views to show Regional projects by region
New views at `/projects/regional/:region/in-progress` and
`/projects/regional/:region/completed` to show projects NOT assigned to
Regional caseworker services, filtered by region.

The region parameter may be passed in with either hyphens or underscores.
E.g.
`/projects/regional/west-midlands/in-progress` and
`/projects/regional/west_midlands/in-progress` are synonymous.

We have done this because other URLs use hyphens (e.g. `/in-progress`) and
if users decide to manually type URLs, it makes sense to allow them to use
hyphens.

If the region parameter is not recognised, the page will return a Not Found
error, e.g. `/projects/regional/paris/in-progress`